# hb_back/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HB Service",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/transaction": {
      "put": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Редактирование транзакции",
        "operationId": "updateTransactionUsingPUT",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«TransactionDto»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Добавление транзакции к пользователю",
        "operationId": "addTransactionUsingPOST",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«TransactionDto»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Удаление транзакции по id",
        "operationId": "deleteTransactionUsingDELETE",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "transactionId",
            "in": "query",
            "description": "Идентификато транзакции",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«int»"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/transaction/balance": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Получение баланса пользователя",
        "operationId": "getBalanceUsingGET",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«bigdecimal»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/transaction/byCategory": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Получение списка транзакций пользователя в определённой категории",
        "operationId": "getByCategoryUsingPOST",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionByCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«PageDto«TransactionDto»»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/transaction/currMonthOutSum": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Получение суммы расходов на все категории за текущий месяц",
        "operationId": "getCurrMonthOutSumUsingGET",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«bigdecimal»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/transaction/list": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Получение списка транзакций пользователя",
        "operationId": "getListUsingPOST",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«PageDto«TransactionDto»»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/transaction/{transactionId}": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Получение транзакции по id",
        "operationId": "getTransactionByIdUsingGET",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false
          },
          {
            "name": "transactionId",
            "in": "path",
            "description": "Идентификато транзакции",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HBResponseData«TransactionDto»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HBResponseData«PageDto«TransactionDto»»": {
        "title": "HBResponseData«PageDto«TransactionDto»»",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Код ответа"
          },
          "data": {
            "$ref": "#/components/schemas/PageDto«TransactionDto»"
          },
          "errorMessage": {
            "type": "string",
            "description": "Ошибка"
          }
        },
        "description": "Ответ HBService"
      },
      "HBResponseData«TransactionDto»": {
        "title": "HBResponseData«TransactionDto»",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Код ответа"
          },
          "data": {
            "$ref": "#/components/schemas/TransactionDto"
          },
          "errorMessage": {
            "type": "string",
            "description": "Ошибка"
          }
        },
        "description": "Ответ HBService"
      },
...
      "HBResponseData«bigdecimal»": {
        "title": "HBResponseData«bigdecimal»",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Код ответа"
          },
          "data": {
            "type": "number",
            "description": "Тело ответа",
            "format": "bigdecimal"
          },
          "errorMessage": {
            "type": "string",
            "description": "Ошибка"
          }
        },
        "description": "Ответ HBService"
      },
      "HBResponseData«int»": {
        "title": "HBResponseData«int»",
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Код ответа"
          },
          "data": {
            "type": "integer",
            "description": "Тело ответа",
            "format": "int32"
          },
          "errorMessage": {
            "type": "string",
            "description": "Ошибка"
          }
        },
        "description": "Ответ HBService"
      },
      "PageDto«TransactionDto»": {
        "title": "PageDto«TransactionDto»",
        "type": "object",
        "properties": {
          "currPage": {
            "type": "integer",
            "description": "Текущая страница",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean",
            "description": "Признак существует ли следующая страница",
            "example": false
          },
          "items": {
            "type": "array",
            "description": "Список элементов",
            "items": {
              "$ref": "#/components/schemas/TransactionDto"
            }
          },
          "limit": {
            "type": "integer",
            "description": "Количество элементов на одной странице",
            "format": "int32"
          },
          "maxPage": {
            "type": "integer",
            "description": "Общее число страниц",
            "format": "int32"
          },
          "sortField": {
            "type": "string",
            "description": "Поле, по которому отсортированы элемента"
          },
          "sortOrder": {
            "type": "string",
            "description": "Порядок сортировки",
            "enum": [
              "ASC",
              "DESC"
            ]
          }
        },
        "description": "Ответ со списком элементов с пагинацией "
      },
      "PageRequest": {
        "title": "PageRequest",
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "description": "Текущая страница",
            "format": "int32"
          },
          "sortField": {
            "type": "string",
            "description": "Поле по которому будут отсортированы элементы"
          },
          "sortOrder": {
            "type": "string",
            "description": "Порядок сортировки",
            "enum": [
              "ASC",
              "DESC"
            ]
          }
        },
        "description": "Пагинация"
      },
      "TransactionByCategoryRequest": {
        "title": "TransactionByCategoryRequest",
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "Идентификатор категории",
            "format": "uuid"
          },
          "page": {
            "$ref": "#/components/schemas/PageRequest"
          },
          "userEmail": {
            "type": "string",
            "description": "Email пользователя"
          }
        },
        "description": "Запрос списка транзакций по категории"
      },
      "TransactionDto": {
        "title": "TransactionDto",
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "Идентификатор категории, на которую была совершена трата(null в случае пополнения)",
            "format": "uuid"
          },
          "createTime": {
            "type": "string",
            "description": "Время транзакции",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "Описание транзакции"
          },
          "sum": {
            "type": "number",
            "description": "Сумма транзакции",
            "format": "bigdecimal"
          },
          "transactionId": {
            "type": "string",
            "description": "Идентификатор транзакции",
            "format": "uuid"
          },
          "userEmail": {
            "type": "string",
            "description": "Email пользователя"
          }
        },
        "description": "Транзакция(пополнение или расход)"
      },
      "TransactionListRequest": {
        "title": "TransactionListRequest",
        "type": "object",
        "properties": {
          "page": {
            "$ref": "#/components/schemas/PageRequest"
          },
          "transactionType": {
            "type": "string",
            "description": "Тип транзакции(пополнение, расход ил все)",
            "enum": [
              "ALL",
              "IN",
              "OUT"
            ]
          },
          "userEmail": {
            "type": "string",
            "description": "Email пользователя"
          }
        },
        "description": "Запрос списка транзакций пользователя"
      },
...
    }
  }
}